fix: 체력 연동 오류 수정

+++ Assets/02.Scripts/03.Player/Entity/EquipmentController.cs
@@ -60,7 +60,7 @@ public class EquipmentController : MonoBehaviour
                 break;
 
             case StatType.Defense:
-                statHandler.AddDefense(modifier.value);
+                statHandler.AddDefense((int)modifier.value);
                 break;
 
             case StatType.MoveSpeedPercent:
